change to call move function instead of reading table context data (#147)
* change to call move function instead of reading table context data

* lint

* change to do not receive module_ids
