+++ crates/e2e-move-tests/src/harness.rs
@@ -110,8 +110,8 @@ impl MoveHarness {
         path: &str,
         upgrade_policy: u8,
     ) -> Result<MessageOutput, VMStatus> {
-        let (module_ids, code) = self.compile_package(path);
-        let msg = self.create_publish_message(*acc, module_ids, code, upgrade_policy);
+        let code = self.compile_package(path);
+        let msg = self.create_publish_message(*acc, code, upgrade_policy);
         self.run_message(msg)
     }
 
@@ -187,7 +187,7 @@ impl MoveHarness {
         vals
     }
 
-    pub fn compile_package(&mut self, path: &str) -> (Vec<String>, Vec<Vec<u8>>) {
+    pub fn compile_package(&mut self, path: &str) -> Vec<Vec<u8>> {
         let package_path = path_in_crate(path);
         let package = BuiltPackage::build(
             package_path.clone(),
@@ -201,27 +201,19 @@ impl MoveHarness {
         )
         .expect("compile failed");
 
-        (
-            package
-                .modules()
-                .map(|v| v.self_id().short_str_lossless())
-                .collect(),
-            package.extract_code(),
-        )
+        package.extract_code()
     }
 
     pub fn create_publish_message(
         &mut self,
         sender: AccountAddress,
-        module_ids: Vec<String>,
         modules: Vec<Vec<u8>>,
         upgrade_policy: u8,
     ) -> Message {
         let ef = MoveHarness::create_entry_function_with_json(
-            str::parse("0x1::code::publish").unwrap(),
+            str::parse("0x1::code::publish_v2").unwrap(),
             vec![],
             vec![
-                serde_json::to_string(&module_ids).unwrap(),
                 serde_json::to_string(&modules.iter().map(hex::encode).collect::<Vec<String>>())
                     .unwrap(),
                 serde_json::to_string(&upgrade_policy).unwrap(), // compatible upgrade policy
