+++ crates/e2e-move-tests/src/tests/cache.rs
@@ -31,12 +31,7 @@ fn test_redeploy_should_update_module_cache() {
     h.initialize();
 
     let code = hex::decode("a11ceb0b060000000d01000a020a12031c28044404054822076aa001088a022010aa02420aec02140b8003020c8203670de903020eeb030200000001000200030004000509010001000600000007060003110700000800010100000900020100000a03040100000b04010002120700000413040801000114050401060505060901050103010b00010900020c03000109000205070b0001090001060c010803010802094261736963436f696e056576656e74067369676e657206737472696e6709747970655f696e666f04436f696e06496e69746961094d696e744576656e7403676574086765745f636f696e046d696e74066e756d6265720576616c756504746573740b64756d6d795f6669656c6406616d6f756e7409636f696e5f7479706506537472696e670a616464726573735f6f6609747970655f6e616d6504656d6974000000000000000000000000000000000000000000000000000000000000000113696e697469613a3a6d657461646174615f76302d0001094d696e744576656e740104000303676574010100066e756d626572010100086765745f636f696e0101000002020c030d010102010e010202020f031008030005000100010004050b003d0037001402010100010004040b003d0014020201040100061d0e0011040c020a023b0020040d0e000a010839003f0005180b023c000c030a033700140a01160b033600150b0138001202380102030100000402064101000000000000020000000500").expect("ms");
-    let msg = h.create_publish_message(
-        AccountAddress::ONE,
-        vec!["0x1::BasicCoin".to_string()],
-        vec![code],
-        1,
-    );
+    let msg = h.create_publish_message(AccountAddress::ONE, vec![code], 1);
     let output = h.run_message(msg).expect("should success");
 
     // commit to store for republish check,
