+++ crates/e2e-move-tests/src/tests/string_viewer.data/viewer/Move.toml
@@ -7,3 +7,4 @@ InitiaStdlib = { local = "../../../../../../precompile/modules/initia_stdlib" }
 
 [addresses]
 std =  "0x1"
+publisher = "0x9999"
