+++ crates/natives/src/code.rs
@@ -1,5 +1,4 @@
 use crate::{
-    helpers::get_string,
     interface::{
         RawSafeNative, SafeNativeBuilder, SafeNativeContext, SafeNativeError, SafeNativeResult,
     },
@@ -9,10 +8,7 @@ use better_any::{Tid, TidAble};
 use initia_move_types::module::ModuleBundle;
 use move_core_types::{account_address::AccountAddress, gas_algebra::NumBytes};
 use move_vm_runtime::native_functions::NativeFunction;
-use move_vm_types::{
-    loaded_data::runtime_types::Type,
-    values::{Struct, Value},
-};
+use move_vm_types::{loaded_data::runtime_types::Type, values::Value};
 use serde::{Deserialize, Serialize};
 use smallvec::{smallvec, SmallVec};
 use std::collections::VecDeque;
@@ -63,7 +59,6 @@ const ECATEGORY_INVALID_ARGUMENT: u64 = 0x1;
 
 // native errors always start from 100
 const EALREADY_REQUESTED: u64 = (ECATEGORY_INVALID_ARGUMENT << 16) + 100;
-const EUNABLE_TO_PARSE_STRING: u64 = (ECATEGORY_INVALID_ARGUMENT << 16) + 101;
 
 /// The native code context.
 #[derive(Tid, Default)]
@@ -76,9 +71,9 @@ pub struct NativeCodeContext {
 /// Represents a request for code publishing made from a native call and to be processed
 /// by the Initia VM.
 pub struct PublishRequest {
-    pub destination: AccountAddress,
+    pub publisher: AccountAddress,
     pub module_bundle: ModuleBundle,
-    pub expected_modules: Option<Vec<String>>,
+    pub upgrade_policy: u8,
 }
 
 /***************************************************************************************************
@@ -112,6 +107,8 @@ fn native_request_publish(
 
     context.charge(gas_params.code_request_publish_base_cost)?;
 
+    let upgrade_policy = safely_pop_arg!(arguments, u8);
+
     let mut code: Vec<Vec<u8>> = vec![];
     for module_code in safely_pop_vec_arg!(arguments, Vec<u8>) {
         context.charge(
@@ -120,21 +117,7 @@ fn native_request_publish(
         code.push(module_code);
     }
 
-    let mut expected_modules: Vec<String> = vec![];
-    for name in safely_pop_vec_arg!(arguments, Struct) {
-        let str_bytes = get_string(name)?;
-
-        context.charge(
-            gas_params.code_request_publish_per_byte * NumBytes::new(str_bytes.len() as u64),
-        )?;
-        expected_modules.push(String::from_utf8(str_bytes).map_err(|_| {
-            SafeNativeError::Abort {
-                abort_code: EUNABLE_TO_PARSE_STRING,
-            }
-        })?);
-    }
-
-    let destination = safely_pop_arg!(arguments, AccountAddress);
+    let publisher = safely_pop_arg!(arguments, AccountAddress);
 
     let code_context = context.extensions_mut().get_mut::<NativeCodeContext>();
     if code_context.requested_module_bundle.is_some() {
@@ -144,9 +127,9 @@ fn native_request_publish(
     }
 
     code_context.requested_module_bundle = Some(PublishRequest {
-        destination,
+        publisher,
         module_bundle: ModuleBundle::new(code),
-        expected_modules: Some(expected_modules),
+        upgrade_policy,
     });
 
     Ok(smallvec![])
