+++ crates/vm/src/lib.rs
@@ -1,6 +1,6 @@
-mod session;
-
 pub use crate::initia_vm::InitiaVM;
 
 mod initia_vm;
+mod publish;
+mod session;
 mod verifier;
